refactor InputPrompt width calculations
